requirements.txt
numpy==1.17.4
Pillow==10.3.0
